# stormborne-booking-backend/docs/swagger.yaml
basePath: /
definitions:
...
  error.AppError:
    properties:
      code:
        type: string
      error: {}
      message:
        type: string
    type: object
...
  model.Movie:
    properties:
      duration:
        type: string
      genre:
        type: string
      id:
        type: string
      imdbRating:
        type: string
      imdbVotes:
        type: string
      name:
        type: string
      plot:
        type: string
      poster:
        type: string
      rated:
        type: string
    type: object
...
  request.OTPResendRequest:
    properties:
      recipient:
        type: string
    required:
    - recipient
    type: object
  request.OTPSendRequest:
    properties:
      recipient:
        type: string
    required:
    - recipient
    type: object
  request.OTPVerifyRequest:
    properties:
      code:
        type: string
      recipient:
        type: string
    required:
    - code
    - recipient
    type: object
...
  response.MovieResponse:
    properties:
      Actors:
        type: string
      Awards:
        type: string
      Country:
        type: string
      Director:
        type: string
      Genre:
        type: string
      Language:
        type: string
      Plot:
        type: string
      Poster:
        type: string
      Rated:
        type: string
      Released:
        type: string
      Runtime:
        type: string
      Title:
        type: string
      Writer:
        type: string
      Year:
        type: string
      imdbID:
        type: string
    type: object
  response.OTPItem:
    properties:
      attempts:
        type: integer
      code:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      is_used:
        type: boolean
      purpose:
        type: string
      recipient:
        type: string
      type:
        type: string
    type: object
  response.OTPListResponse:
    properties:
      otps:
        items:
          $ref: '#/definitions/response.OTPItem'
        type: array
    type: object
  response.OTPSendResponse:
    properties:
      expires_in_seconds:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  response.OTPVerifyResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      verified:
        type: boolean
    type: object
...
  response.SeatResponse:
    properties:
      basePrice:
        description: includes weekend surcharge if applicable
        type: number
      column:
        description: e.g. 1
        type: integer
      isWeekend:
        type: boolean
      label:
        description: e.g. "A1", "B10"
        type: string
      row:
        description: e.g. "A"
        type: string
      seatId:
        type: integer
      seatType:
        description: '"REGULAR", "PREMIUM"'
        type: string
      status:
        description: '"available" | "sold"'
        type: string
    type: object
  response.SeatStatusResponse:
    properties:
      seats:
        items:
          $ref: '#/definitions/response.SeatResponse'
        type: array
      showDate:
        type: string
      showId:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a skyfox
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Booking API
  version: "1.0"
paths:
  /admin/movies/{movieId}:
    get:
      consumes:
      - application/json
      description: get movie details by movie id
      parameters:
      - description: Enter basic auth
        in: header
        name: Authorization
        required: true
        type: string
      - description: movie id
        in: path
        name: movieId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Movie'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BasicAuth: []
      summary: Movie Details
      tags:
      - Movies
...
  /movies:
    get:
      consumes:
      - application/json
      description: get shows by date
      parameters:
      - description: Enter basic auth
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.MovieResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BasicAuth: []
      summary: Movie
      tags:
      - Movies
...
  /otp/all:
    get:
      description: Returns all OTPs for the illusion dashboard (no auth required)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OTPListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      summary: List All OTPs
      tags:
      - OTP
  /otp/email/request:
    post:
      consumes:
      - application/json
      description: Generate and send OTP to an email address (illusion mode)
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.OTPSendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OTPSendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Request Email OTP
      tags:
      - OTP
  /otp/email/resend:
    post:
      consumes:
      - application/json
      description: Resend OTP to an email address with rate limiting
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.OTPResendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OTPSendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Resend Email OTP
      tags:
      - OTP
  /otp/email/verify:
    post:
      consumes:
      - application/json
      description: Verify the OTP code sent to an email address
      parameters:
      - description: Email address and OTP code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.OTPVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OTPVerifyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Verify Email OTP
      tags:
      - OTP
  /otp/sms/request:
    post:
      consumes:
      - application/json
      description: Generate and send OTP to a phone number (illusion mode)
      parameters:
      - description: Phone number
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.OTPSendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OTPSendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Request SMS OTP
      tags:
      - OTP
  /otp/sms/resend:
    post:
      consumes:
      - application/json
      description: Resend OTP to a phone number with rate limiting
      parameters:
      - description: Phone number
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.OTPResendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OTPSendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Resend SMS OTP
      tags:
      - OTP
  /otp/sms/verify:
    post:
      consumes:
      - application/json
      description: Verify the OTP code sent to a phone number
      parameters:
      - description: Phone number and OTP code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.OTPVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OTPVerifyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Verify SMS OTP
      tags:
      - OTP
...
  /shows/{showId}/seats:
    get:
      consumes:
      - application/json
      description: Returns all seats with availability status and pricing for a given
        show. Weekend shows (Sat/Sun) have a ₹50 surcharge applied to base price.
      parameters:
      - description: Show ID
        example: 1
        in: path
        name: showId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SeatStatusResponse'
        "400":
          description: Invalid showId
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Feature not available
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get seat map for a show
      tags:
      - seats
...
swagger: "2.0"